requirements.txt
aiogram==3.0.0rc2
aiohttp==3.9.5
python-dotenv==1.1.1
nest_asyncio==1.6.0
